text_preprocessing/app/requirements.txt
Flask==2.1.1
requests==2.26.0
PyYAML==6.0
transformers==4.14.1
